# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Beautiful Island Tours API Documentation",
        "description": "API documentation for Beautiful Island Tours system",
        "contact": {
            "name": "Beautiful Island Tours Support",
            "email": "ashanhimantha321@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://testapi.ashanhimantha.com",
            "description": "Beautiful Island Tours API Server"
        }
    ],
    "paths": {
...
        "/api/taxis": {
            "get": {
                "tags": [
                    "Taxis"
                ],
                "summary": "Get list of taxis",
                "description": "Display a listing of taxis.",
                "operationId": "e95d9387e56a40e9828c11525ca32616",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter taxis by status",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "maintenance",
                                "in_tour"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Taxi"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Taxis"
                ],
                "summary": "Create new taxi",
                "description": "Store a newly created taxi in storage.",
                "operationId": "f2ac4c84aa517493635a1980a52e2b25",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "engine_capacity",
                                    "kmpl",
                                    "fuel_type",
                                    "gear_type",
                                    "passenger_count",
                                    "cost_per_day",
                                    "description",
                                    "status",
                                    "display_image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Toyota Corolla"
                                    },
                                    "engine_capacity": {
                                        "type": "string",
                                        "example": "1800cc"
                                    },
                                    "kmpl": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 14.5
                                    },
                                    "fuel_type": {
                                        "type": "string",
                                        "example": "Petrol"
                                    },
                                    "gear_type": {
                                        "type": "string",
                                        "example": "Automatic"
                                    },
                                    "passenger_count": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "cost_per_day": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 50
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Comfortable sedan for family trips"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance",
                                            "in_tour"
                                        ],
                                        "example": "active"
                                    },
                                    "display_image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image1": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image2": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "image3": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Taxi created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Taxi created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Taxi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/taxis/{id}/status": {
            "put": {
                "tags": [
                    "Taxis"
                ],
                "summary": "Update taxi status",
                "description": "Update the status of a taxi.",
                "operationId": "9b3320f5e114fc971702581d3d921a9a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Taxi ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "maintenance",
                                            "in_tour"
                                        ],
                                        "example": "active"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Taxi status updated successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Taxi"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Taxi not found"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Taxi": {
                "required": [
                    "title",
                    "engine_capacity",
                    "kmpl",
                    "fuel_type",
                    "gear_type",
                    "passenger_count",
                    "cost_per_day"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "type": "string"
                    },
                    "engine_capacity": {
                        "type": "string"
                    },
                    "kmpl": {
                        "type": "number",
                        "format": "float"
                    },
                    "fuel_type": {
                        "type": "string"
                    },
                    "gear_type": {
                        "type": "string"
                    },
                    "passenger_count": {
                        "type": "integer"
                    },
                    "cost_per_day": {
                        "type": "number",
                        "format": "float"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "maintenance",
                            "in_tour"
                        ]
                    },
                    "display_image": {
                        "type": "string"
                    },
                    "image1": {
                        "type": "string",
                        "nullable": true
                    },
                    "image2": {
                        "type": "string",
                        "nullable": true
                    },
                    "image3": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}